attribute class refactor
