+++ modules/core/src/index.js
@@ -43,7 +43,7 @@ export {default as _LayersPass} from './passes/layers-pass';
 export {default as Deck} from './lib/deck';
 
 export {default as LayerManager} from './lib/layer-manager';
-export {default as AttributeManager} from './lib/attribute-manager';
+export {default as AttributeManager} from './lib/attribute/attribute-manager';
 export {default as Layer} from './lib/layer';
 export {default as CompositeLayer} from './lib/composite-layer';
 export {default as DeckRenderer} from './lib/deck-renderer';
