+++ test/modules/core/lib/attribute/attribute-manager.spec.js
@@ -19,7 +19,7 @@
 // THE SOFTWARE.
 
 /* eslint-disable dot-notation, max-statements, no-unused-vars */
-import AttributeManager from '@deck.gl/core/lib/attribute-manager';
+import AttributeManager from '@deck.gl/core/lib/attribute/attribute-manager';
 import GL from '@luma.gl/constants';
 import test from 'tape-catch';
 import {gl} from '@deck.gl/test-utils';
@@ -176,7 +176,7 @@ test('AttributeManager.update - external buffers', t => {
     }
   });
 
-  t.is(attribute.type, gl.FLOAT, 'colors accessor is set to correct type');
+  t.is(attribute.shaderAttributes.colors.type, gl.FLOAT, 'colors accessor is set to correct type');
 
   attributeManager.update({
     numInstances: 1,
@@ -185,7 +185,7 @@ test('AttributeManager.update - external buffers', t => {
       colors: new Uint32Array([0, 0, 0])
     }
   });
-  t.is(attribute.type, gl.UNSIGNED_INT, 'colors accessor is set to correct type');
+  t.is(attribute.shaderAttributes.colors.type, gl.UNSIGNED_INT, 'colors accessor is set to correct type');
 
   t.end();
 });
