+++ test/modules/core/lib/attribute/attribute-transition-manager.spec.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-statements */
-import AttributeTransitionManager from '@deck.gl/core/lib/attribute-transition-manager';
-import Attribute from '@deck.gl/core/lib/attribute';
+import AttributeTransitionManager from '@deck.gl/core/lib/attribute/attribute-transition-manager';
+import Attribute from '@deck.gl/core/lib/attribute/attribute';
 import test from 'tape-catch';
 import {isWebGL2} from '@luma.gl/core';
 import {Timeline} from '@luma.gl/addons';
