+++ test/modules/core/lib/attribute/base-attribute.spec.js
