+++ test/modules/core/lib/attribute/index.js
@@ -1,0 +1,4 @@
+// import './base-attribute.spec';
+import './attribute.spec';
+import './attribute-manager.spec';
+import './attribute-transition-manager.spec';
